add Wsign-conversion and fix

+++ framework/include/I2CDevObj.hpp
@@ -58,17 +58,17 @@ public:
 	virtual int start();
 	virtual int stop();
 
-	static int readReg(DevHandle &h, uint8_t address, uint8_t *out_buffer, int length);
-	static int writeReg(DevHandle &h, uint8_t address, uint8_t *in_buffer, int length);
+	static int readReg(DevHandle &h, uint8_t address, uint8_t *out_buffer, size_t length);
+	static int writeReg(DevHandle &h, uint8_t address, uint8_t *in_buffer, size_t length);
 
 protected:
-	int _readReg(uint8_t address, uint8_t *out_buffer, int length);
-	int _writeReg(uint8_t address, uint8_t *out_buffer, int length);
-	int _readReg16(uint16_t address, uint16_t *out_buffer, int length);
-	int _writeReg16(uint16_t address, uint16_t *out_buffer, int length);
+	int _readReg(uint8_t address, uint8_t *out_buffer, size_t length);
+	int _writeReg(uint8_t address, uint8_t *out_buffer, size_t length);
+	int _readReg16(uint16_t address, uint16_t *out_buffer, size_t length);
+	int _writeReg16(uint16_t address, uint16_t *out_buffer, size_t length);
 
 	// read from a register without ioctl
-	int _simple_read(uint8_t *out_buffer, int length);
+	int _simple_read(uint8_t *out_buffer, size_t length);
 
 	int _setSlaveConfig(uint32_t slave_address, uint32_t bus_frequency_khz, uint32_t transfer_timeout_usec);
